+++ functions/_adapters/golemio/services/vehicles/colors.ts
@@ -1,7 +1,7 @@
 /**
  * Official PID Branding Colors
  */
-export const VEHICLE_COLORS = {
+const VEHICLE_COLORS = {
     METRO_A: '#00A562',
     METRO_B: '#F8B322',
     METRO_C: '#CF003D',